filter hotels by city name
